pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.35"
faker = "^30.0.0"
factory-boy = "^3.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

